+++ R/plans.R
@@ -12,12 +12,14 @@
 #' * `max_year`: The maximum year to be analyzed, supplied in the `max_year` argument.
 #' * `iea_data_path`: The path to IEA extended energy balance data, supplied in the `iea_data_path` argument.
 #' * `ceda_data_folder`: The path to the CEDA data, supplied in the `ceda_data_folder` argument.
+#' * `machine_data_path`: The path to the machine data excel files, supplied in the `machine_data_path` argument.
 #' * `exemplar_table_path`: The path to an exemplar table, supplied in the `exemplar_table_path` argument.
 #' * `fu_analysis_folder`: The path to the final-to-useful analysis folder, supplied in the `fu_analysis_folder` argument.
 #' * `report_output_folder`: The path to a report output folder, supplied in the `report_output_folder` argument.
 #' * `AllIEAData`: A data frame with all IEA extended energy balance data read from `iea_data_path`.
 #' * `IEAData`: A version of the `AllIEAData` data frame containing data for only those countries specified in `countries`.
 #' * `CEDAData`: A data frame containing temperature data supplied through `CEDATools::read_cru_cy_files`.
+#' * `AllMachineData`: A data frame containing Eta.fu and Phi.u values read through functions in `machine_functions.R`.
 #' * `balanced_before`: A boolean that tells where the data were balanced as received, usually a vector of `FALSE`, one for each country.
 #' * `BalancedIEAData`: A data frame containing balanced IEA extended energy balance data.
 #' * `balanced_after`: A boolean telling whether IEA extended energy balance data is balanced after balancing, usually a vector of `TRUE`, one for each country.
@@ -69,6 +71,7 @@
 #'                Default is "all_targets" to indicate all targets of the plan should be returned.
 #' @param iea_data_path The path to IEA extended energy balance data in .csv format.
 #' @param ceda_data_folder The path to the CEDA data in text file, .per, format.
+#' @param machine_data_path The path to the machine data in .xlsx format.
 #' @param exemplar_table_path The path to an exemplar table.
 #' @param fu_analysis_folder The path to a folder containing final-to-useful analyses.
 #'                           Sub-folders named with 3-letter country abbreviations are assumed.
@@ -92,13 +95,15 @@
 #'          max_year = 1999,
 #'          iea_data_path = "iea_path",
 #'          ceda_data_folder = "ceda_path",
+#'          machine_data_path = "machine_path",
 #'          exemplar_table_path = "exemplar_path",
 #'          fu_analysis_folder = "fu_folder",
 #'          reports_source_folders = "reports_source_folders",
 #'          reports_dest_folder = "reports_dest_folder")
 get_plan <- function(countries, additional_exemplar_countries = NULL,
                      max_year, how_far = "all_targets",
-                     iea_data_path, ceda_data_folder, exemplar_table_path,
+                     iea_data_path, ceda_data_folder,
+                     machine_data_path, exemplar_table_path,
                      fu_analysis_folder, reports_source_folders, reports_dest_folder) {
 
   # Get around warnings of type "no visible binding for global variable".
@@ -107,6 +112,8 @@ get_plan <- function(countries, additional_exemplar_countries = NULL,
   AllIEAData <- NULL
   IEAData <- NULL
   CEDAData <- NULL
+  AllMachineData <- NULL
+  MachineData <- NULL
   BalancedIEAData <- NULL
   balanced_after <- NULL
   Specified <- NULL
@@ -131,6 +138,7 @@ get_plan <- function(countries, additional_exemplar_countries = NULL,
     max_year = !!max_year,
     iea_data_path = !!iea_data_path,
     ceda_data_folder = !!ceda_data_folder,
+    machine_data_path = !!machine_data_path,
     exemplar_table_path = !!exemplar_table_path,
     fu_analysis_folder = !!fu_analysis_folder,
     reports_source_folders = !!reports_source_folders,
@@ -143,11 +151,15 @@ get_plan <- function(countries, additional_exemplar_countries = NULL,
                               extract_country_data(countries = alloc_and_eff_couns, max_year = max_year),
                             dynamic = map(alloc_and_eff_couns)),
 
-    # (1a) Grab all CEDA data for ALL countries
+    # (1b) Grab all CEDA data for ALL countries
 
     CEDAData = drake::target(CEDATools::create_agg_cru_cy_df(agg_cru_cy_folder = ceda_data_folder,
                                                              agg_cru_cy_metric = c("tmp", "tmn", "tmx"),
                                                              agg_cru_cy_year = 2020)),
+    # (1c) Grab Machine data for ALL countries
+
+    AllMachineData = drake::target(read_all_eta_files(eta_fin_paths = get_eta_filepaths(machine_data_path))),
+
 
 
     # (2) Balance all final energy data.
@@ -188,14 +200,14 @@ get_plan <- function(countries, additional_exemplar_countries = NULL,
                                     load_exemplar_table(countries = alloc_and_eff_couns,
                                                         max_year = max_year) %>%
                                     exemplar_lists(alloc_and_eff_couns),
-                                  dynamic = map(alloc_and_eff_couns)),
+                                    dynamic = map(alloc_and_eff_couns)),
 
     # (6) Load incomplete FU allocation tables
 
     IncompleteAllocationTables = drake::target(fu_analysis_folder %>%
                                                  load_fu_allocation_tables(specified_iea_data = Specified,
                                                                            countries = alloc_and_eff_couns),
-                                               dynamic = map(alloc_and_eff_couns)),
+                                                 dynamic = map(alloc_and_eff_couns)),
 
     # (7) Complete FU allocation tables