+++ NEWS.md
@@ -1,3 +1,11 @@
+* New target "AllMachineData" which consists of a tidy data 
+  frames containing information obtained by calling functions in 
+  machine_functions.R.
+
+* New set of functions in machine_functions.R (`get_eta_filepaths()` and 
+  `read_all_eta_files`) which read in data from a specified file path and create 
+  a tidy data frame containing Eta.fu, Phi.u, and any additional metrics supplied.
+
 * New version of plan that includes extending the PSUT matrices to the useful stage.
 * New functions `nonstat_alloc_graph()` and `nonstat_alloc_plots_df()`
   which make final-to-useful allocations graphs for only non-stationary allocations